Move chapter 4 exercises and solutions into standard layout.
