+++ src/test/kotlin/chapter4/exercises/ex2/listing.kt
@@ -1,12 +1,13 @@
-package chapter4.exercises
+package chapter4.exercises.ex2
 
 import chapter3.List
-import chapter4.Boilerplate.isEmpty
-import chapter4.Boilerplate.size
-import chapter4.Boilerplate.sum
+import chapter4.isEmpty
+import chapter4.size
+import chapter4.sum
 import chapter4.None
 import chapter4.Option
 import chapter4.Some
+import chapter4.getOrElse
 import io.kotlintest.matchers.doubles.plusOrMinus
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
@@ -19,7 +20,7 @@ fun mean(xs: List<Double>): Option<Double> =
 fun variance(xs: List<Double>): Option<Double> = TODO()
 //end::init[]
 
-class Exercise_4_2 : WordSpec({
+class Exercise2 : WordSpec({
 
     "variance" should {
         "!determine the variance of a list of numbers" {
