+++ src/test/kotlin/chapter4/exercises/ex3/listing.kt
@@ -1,4 +1,4 @@
-package chapter4.solutions
+package chapter4.exercises.ex3
 
 import chapter4.None
 import chapter4.Option
@@ -7,19 +7,11 @@ import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
 //tag::init[]
-fun <A, B, C> map2(
-    oa: Option<A>,
-    ob: Option<B>,
-    f: (A, B) -> C
-): Option<C> =
-    oa.flatMap { a ->
-        ob.map { b ->
-            f(a, b)
-        }
-    }
+fun <A, B, C> map2(a: Option<A>, b: Option<B>, f: (A, B) -> C): Option<C> =
+    TODO()
 //end::init[]
 
-class Solution_4_3 : WordSpec({
+class Exercise3 : WordSpec({
 
     "map2" should {
 
@@ -27,16 +19,17 @@ class Solution_4_3 : WordSpec({
         val b = Some(20)
         val none = Option.empty<Int>()
 
-        "combine two option values using a binary function" {
+        "!combine two option values using a binary function" {
             map2(a, b) { aa, bb ->
                 aa * bb
             } shouldBe Some(100)
         }
 
-        "return none if either option is not defined" {
+        "!return none if either option is not defined" {
             map2(a, none) { aa, bb ->
                 aa * bb
             } shouldBe None
+
             map2(none, b) { aa, bb ->
                 aa * bb
             } shouldBe None
