+++ src/test/kotlin/chapter4/solutions/sol4/listing.kt
@@ -1,12 +1,13 @@
-package chapter4.solutions
+package chapter4.solutions.sol4
 
 import chapter3.Cons
 import chapter3.List
 import chapter3.Nil
-import chapter4.Boilerplate.foldRight
+import chapter4.foldRight
 import chapter4.None
 import chapter4.Option
 import chapter4.Some
+import chapter4.map2
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
@@ -22,7 +23,7 @@ fun <A> sequence(
         })
 //end::init[]
 
-class Solution_4_4 : WordSpec({
+class Solution4 : WordSpec({
 
     "sequence" should {
         "turn a list of some options into an option of list" {
